Re-integrate JDBC and JNDI features into duke-core.
(See larsga/Duke#150 for more information)
