+++ duke-core/src/main/java/no/priv/garshol/duke/JDBCEquivalenceClassDatabase.java
@@ -1,16 +1,14 @@
 
 package no.priv.garshol.duke;
 
-import java.util.List;
-import java.util.Iterator;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Properties;
-
-import java.sql.Connection;
-import java.sql.Statement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Properties;
 
 import no.priv.garshol.duke.utils.JDBCUtils;
