+++ duke-core/src/main/java/no/priv/garshol/duke/JNDILinkDatabase.java
