+++ duke-core/src/main/java/no/priv/garshol/duke/RDBMSLinkDatabase.java
@@ -1,15 +1,14 @@
 
 package no.priv.garshol.duke;
 
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Properties;
-import java.util.Collection;
 import java.sql.Connection;
-import java.sql.Statement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
 import no.priv.garshol.duke.utils.JDBCUtils;
