+++ duke-core/src/main/java/no/priv/garshol/duke/datasources/JDBCDataSource.java
