+++ duke-core/src/main/java/no/priv/garshol/duke/utils/JDBCUtils.java
@@ -1,17 +1,16 @@
 
 package no.priv.garshol.duke.utils;
 
-import java.util.Properties;
-import java.sql.Driver;
-import java.sql.Statement;
-import java.sql.ResultSet;
-import java.sql.Connection;
-import java.sql.SQLException;
-
 import javax.naming.Context;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.Driver;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.Properties;
 
 import no.priv.garshol.duke.DukeException;
