+++ duke-core/src/test/java/no/priv/garshol/duke/JDBCClassDatabaseTest.java
