+++ duke-core/src/test/java/no/priv/garshol/duke/JDBCLinkDatabaseTest.java
