+++ duke-core/src/test/java/no/priv/garshol/duke/matchers/InMemoryLinkDatabaseMatchListenerTest.java
