+++ duke-core/src/test/java/no/priv/garshol/duke/matchers/LinkDatabaseMatchListenerTest.java
