--- duke-jdbc/pom.xml
@@ -1,26 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>no.priv.garshol.duke</groupId>
-        <artifactId>duke</artifactId>
-        <version>1.3-SNAPSHOT</version>
-        <relativePath>../</relativePath>
-    </parent>
-    <artifactId>duke-jdbc</artifactId>
-    <packaging>jar</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>no.priv.garshol.duke</groupId>
-            <artifactId>duke-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>no.priv.garshol.duke</groupId>
-            <artifactId>duke-core</artifactId>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
-
-    </dependencies>
-</project>
