+++ duke-server/pom.xml
@@ -22,10 +22,6 @@
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>no.priv.garshol.duke</groupId>
-            <artifactId>duke-jdbc</artifactId>
-        </dependency>
 
         <!-- ===== SERVER DEPENDENCIES -->
